requirements.txt
requests
langchain
langchain-community
python-docx
pypdf
tqdm
python-dotenv
docx2txt